pyproject.toml
[tool.poetry]
name = "django-mediastorage"
version = "0.1.0"
description = ""
authors = ["alissa <alissa.gerhard@jonasundderwolf.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12"
django = ">=4.2"
djangorestframework = ">=3.15"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "*"
pytest-factoryboy = "*"
pytest-django = "*"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pre-commit = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 90

[tool.ruff.lint]
ignore = [
    "E501", # pycodestyle: line-too-long
    "T201", # flake8-print: print statement found
]
select = [
    "B0",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "E",   # pycodestyle (Error)
    "F",   # Pyflakes
    "I",   # isort
    "T20", # flake8-print
    "W",   # pycodestyle (Warning)
]

[tool.ruff.lint.isort]
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "project",
    "local-folder",
]

[tool.ruff.lint.isort.sections]
"project" = ["django_mediastorage"]

[tool.ruff.lint.mccabe]
max-complexity = 9
